Fix: Update .gitignore to ignore more files
This commit updates the `.gitignore` file to ignore more files that are generally not necessary to track in version control. This includes files generated by various tools and IDEs, as well as temporary files and build artifacts.

The goal is to reduce the size of the repository and make it easier to manage.

This addresses issues related to large repository size and inconsistent file inclusion.